+++ base/rico-client/src/main/java/dev/rico/internal/client/http/HttpCallRequestBuilderImpl.java
@@ -16,14 +16,11 @@
  */
 package dev.rico.internal.client.http;
 
+import dev.rico.core.http.*;
 import dev.rico.internal.core.Assert;
 import dev.rico.internal.core.http.HttpClientConnection;
 import dev.rico.internal.core.http.HttpHeaderImpl;
 import dev.rico.client.ClientConfiguration;
-import dev.rico.core.http.ByteArrayProvider;
-import dev.rico.core.http.HttpCallRequestBuilder;
-import dev.rico.core.http.HttpCallResponseBuilder;
-import dev.rico.core.http.HttpURLConnectionHandler;
 import com.google.gson.Gson;
 import org.apiguardian.api.API;
 
@@ -36,63 +33,17 @@ import static dev.rico.internal.core.http.HttpHeaderConstants.JSON_MIME_TYPE;
 import static org.apiguardian.api.API.Status.INTERNAL;
 
 @API(since = "0.x", status = INTERNAL)
-public class HttpCallRequestBuilderImpl implements HttpCallRequestBuilder {
-
-    private final HttpClientConnection connection;
-
-    private final Gson gson;
-
-    private final List<HttpURLConnectionHandler> requestHandlers;
-
-    private final List<HttpURLConnectionHandler> responseHandlers;
+public class HttpCallRequestBuilderImpl extends AbstractHttpCallRequestBuilder {
 
     private final ClientConfiguration configuration;
 
-    private final AtomicBoolean done = new AtomicBoolean(false);
-
     public HttpCallRequestBuilderImpl(final HttpClientConnection connection, final Gson gson, final List<HttpURLConnectionHandler> requestHandlers, final List<HttpURLConnectionHandler> responseHandlers, ClientConfiguration configuration) {
-        this.connection = Assert.requireNonNull(connection, "connection");
-        this.gson = Assert.requireNonNull(gson, "gson");
+        super(connection, gson, requestHandlers, responseHandlers);
         this.configuration = configuration;
-
-
-        Assert.requireNonNull(requestHandlers, "requestHandlers");
-        this.requestHandlers = Collections.unmodifiableList(requestHandlers);
-
-        Assert.requireNonNull(responseHandlers, "responseHandlers");
-        this.responseHandlers = Collections.unmodifiableList(responseHandlers);
     }
 
     @Override
-    public HttpCallRequestBuilder withHeader(final String name, final String content) {
-        connection.addRequestHeader(new HttpHeaderImpl(name, content));
-        return this;
-    }
-
-    @Override
-    public HttpCallResponseBuilder withContent(final byte[] content, final String contentType) {
-        withHeader(CONTENT_TYPE_HEADER, contentType);
-        connection.setDoOutput(true);
-        return continueWithResponseBuilder(() -> content);
-    }
-
-    @Override
-    public <I> HttpCallResponseBuilder withContent(final I content) {
-        return withContent(gson.toJson(content), JSON_MIME_TYPE);
-    }
-
-    @Override
-    public HttpCallResponseBuilder withoutContent() {
-        return continueWithResponseBuilder(() -> new byte[0]);
-    }
-
-    private HttpCallResponseBuilder continueWithResponseBuilder(final ByteArrayProvider dataProvider) {
-        Assert.requireNonNull(dataProvider, "dataProvider");
-        if (done.get()) {
-            throw new RuntimeException("Request already defined!");
-        }
-        done.set(true);
+    protected HttpCallResponseBuilder createResponseBuilder(HttpClientConnection connection, ByteArrayProvider dataProvider, Gson gson, List<HttpURLConnectionHandler> requestHandlers, List<HttpURLConnectionHandler> responseHandlers) {
         return new HttpCallResponseBuilderImpl(connection, dataProvider, gson, requestHandlers, responseHandlers, configuration);
     }
-
 }
